+++ Firmware/Projects/MCU_App.X/IO.h
@@ -136,18 +136,18 @@ uint16_t IO_GET_CURRENT_CHARGE_CONTROLLER(void);
 uint16_t IO_GET_CURRENT_MOTOR_CONTROLLER(void);
 uint16_t IO_GET_CURRENT_BMS_CONTROLLER(void);
 uint16_t IO_GET_CURRENT_SPARE_1_CONTROLLER(void);
-uint16_t IO_GET_CURRENT_BATT(void);
-uint16_t IO_GET_CURRENT_DCDC(void);
-uint16_t IO_GET_CURRENT_IC_CONTROLLER(void);
+float IO_GET_CURRENT_BATT(void);
+float IO_GET_CURRENT_DCDC(void);
+float IO_GET_CURRENT_IC_CONTROLLER(void);
 
 /**
  * IO_GET_VOLTAGE_xxx will return the converted voltage (including divider) in milliVolts
  * @return Voltage in milliVolts
  */
-uint16_t IO_GET_VOLTAGE_PILOT(void);
-uint16_t IO_GET_VOLTAGE_PROXIMITY(void);
-uint16_t IO_GET_VOLTAGE_VBAT(void);
-uint16_t IO_GET_VOLTAGE_VBAT_SW(void);
+float IO_GET_VOLTAGE_PILOT(void);
+float IO_GET_VOLTAGE_PROXIMITY(void);
+float IO_GET_VOLTAGE_VBAT(void);
+float IO_GET_VOLTAGE_VBAT_SW(void);
 
 
 /**
